+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
 - [lib] Introduce `QuantumToolboxCore` library. ([#686])
+- [lib] Move more basic functionalities to `QuantumToolboxCore` library. ([#751])
 
 ## [v0.47.3]
 Release date: 2026-07-28
@@ -562,3 +563,4 @@ Release date: 2024-11-13
 [#745]: https://github.com/qutip/QuantumToolbox.jl/issues/745
 [#747]: https://github.com/qutip/QuantumToolbox.jl/issues/747
 [#748]: https://github.com/qutip/QuantumToolbox.jl/issues/748
+[#751]: https://github.com/qutip/QuantumToolbox.jl/issues/751
