+++ lib/QuantumToolboxCore/src/deprecated.jl
@@ -92,13 +92,3 @@ function liouvillian(
     )
     return liouvillian(H, c_ops; kwargs...)
 end
-
-export liouvillian_generalized
-function liouvillian_generalized(args...; kwargs...)
-    Base.depwarn(
-        "`liouvillian_generalized` is deprecated and will be removed in next major release, use `liouvillian_dressed_nonsecular` instead.",
-        :liouvillian_generalized,
-        force = true,
-    )
-    return liouvillian_dressed_nonsecular(args...; kwargs...)
-end
